+++ src/app/components/PokemonList.tsx
@@ -3,6 +3,7 @@ import "./PokemonList.scss";
 
 import { getPokemonList } from "@/server/pokemon/get";
 import Image from "next/image";
+import Link from "next/link";
 import { useEffect } from "react";
 import { useServerAction } from "zsa-react";
 
@@ -26,22 +27,21 @@ const PokemonList = () => {
   return (
     <div className="pokemon-list">
       {pokemonList.map((pokemon) => (
-        <div
-          key={pokemon.id}
-          className={`pokemon ${pokemon.types[0].type.name}`}
-        >
-          <div>
-            <Image
-              src={`https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${pokemon.id}.png`}
-              alt={`Image of ${pokemon.name}`}
-              width={100}
-              height={100}
-            />
+        <Link key={pokemon.id} href={`/pokemon/${pokemon.id}`}>
+          <div className={`pokemon ${pokemon.types[0].type.name}`}>
+            <div>
+              <Image
+                src={`https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${pokemon.id}.png`}
+                alt={`Image of ${pokemon.name}`}
+                width={100}
+                height={100}
+              />
+            </div>
+            <div>{pokemon.name}</div>
+            <div>{pokemon.weight}</div>
+            <div>{pokemon.types.map((type) => type.type.name).join(", ")}</div>
           </div>
-          <div>{pokemon.name}</div>
-          <div>{pokemon.weight}</div>
-          <div>{pokemon.types.map((type) => type.type.name).join(", ")}</div>
-        </div>
+        </Link>
       ))}
     </div>
   );